Modify struct, functions to enable handling message with prev block hash
