+++ core/api_ipc.go
@@ -13,7 +13,7 @@ import (
 
 type RCIPC struct {
 	conn ipc.Connection
-	id uint32
+	id   uint32
 }
 
 func InitRCIPC(net string, address string) (*RCIPC, error) {
@@ -57,7 +57,7 @@ func (rc *RCIPC) SendVersion() (*ResponseVersion, error) {
 	return resp, err
 }
 
-func (rc *RCIPC) SendClaim(address string, blockHeight uint64, blockHash string,
+func (rc *RCIPC) SendClaim(address string, blockHeight uint64, blockHash string, prevBlockHash string,
 	txIndex uint64, txHash string, noCommitClaim bool, noCommitBlock bool) (*ResponseClaim, error) {
 	var req ClaimMessage
 	resp := new(ResponseClaim)
@@ -76,10 +76,21 @@ func (rc *RCIPC) SendClaim(address string, blockHeight uint64, blockHash string,
 		}
 		copy(req.BlockHash, bh)
 	}
+	req.PrevBlockHash = make([]byte, BlockHashSize)
+	if len(prevBlockHash) == 0 {
+		binary.BigEndian.PutUint64(req.PrevBlockHash, blockHeight-1)
+	} else {
+		bh, err := hex.DecodeString(prevBlockHash)
+		if err != nil {
+			log.Printf("Failed to send CLAIM. Invalid prev block hash. %v\n", err)
+			return nil, err
+		}
+		copy(req.PrevBlockHash, bh)
+	}
 	req.TXIndex = txIndex
 	req.TXHash = make([]byte, TXHashSize)
 	if len(txHash) == 0 {
-		binary.BigEndian.PutUint64(req.TXHash, blockHeight + txIndex)
+		binary.BigEndian.PutUint64(req.TXHash, blockHeight+txIndex)
 	} else {
 		th, err := hex.DecodeString(txHash)
 		if err != nil {
@@ -128,7 +139,7 @@ func (rc *RCIPC) SendQuery(address string) (*ResponseQuery, error) {
 	return resp, err
 }
 
-func (rc *RCIPC) SendCalculate(iissData string, blockHeight uint64) (*CalculateResponse, error){
+func (rc *RCIPC) SendCalculate(iissData string, blockHeight uint64) (*CalculateResponse, error) {
 	var req CalculateRequest
 	resp := new(CalculateResponse)
 
@@ -209,7 +220,7 @@ func (rc *RCIPC) SendCommitClaim(success bool, address string, blockHeight uint6
 	req.TXIndex = txIndex
 	req.TXHash = make([]byte, TXHashSize)
 	if len(txHash) == 0 {
-		binary.BigEndian.PutUint64(req.TXHash, blockHeight + txIndex)
+		binary.BigEndian.PutUint64(req.TXHash, blockHeight+txIndex)
 	} else {
 		th, err := hex.DecodeString(txHash)
 		if err != nil {
@@ -227,7 +238,7 @@ func (rc *RCIPC) SendCommitClaim(success bool, address string, blockHeight uint6
 	return err
 }
 
-func (rc *RCIPC) SendQueryCalculateStatus() (*QueryCalculateStatusResponse, error){
+func (rc *RCIPC) SendQueryCalculateStatus() (*QueryCalculateStatusResponse, error) {
 	resp := new(QueryCalculateStatusResponse)
 
 	// Send QUERY_CALCULATE_STATUS and get response
