+++ core/db_precommitInfo.go
@@ -1,0 +1,63 @@
+package core
+
+import (
+	"github.com/icon-project/rewardcalculator/common/codec"
+	"github.com/icon-project/rewardcalculator/common/db"
+	"log"
+)
+
+type PreCommitHierarchy struct {
+	blockHash           [BlockHashSize]byte
+	childrenBlockHashes [][BlockHashSize]byte
+}
+
+func (ph *PreCommitHierarchy) ID() [BlockHashSize]byte {
+	return ph.blockHash
+}
+
+func (ph *PreCommitHierarchy) Bytes() ([]byte, error) {
+	var bytes []byte
+	if bs, err := codec.MarshalToBytes(&ph.childrenBlockHashes); err != nil {
+		return nil, err
+	} else {
+		bytes = bs
+	}
+	return bytes, nil
+}
+
+func (ph *PreCommitHierarchy) SetBytes(bs []byte) error {
+	_, err := codec.UnmarshalFromBytes(bs, &ph.childrenBlockHashes)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+type PreCommitInfo map[[BlockHashSize]byte]map[[BlockHashSize]byte]bool
+
+func LoadPreCommitInfo(pdb db.Database) (preCommitInfo PreCommitInfo) {
+	preCommitInfo = make(PreCommitInfo, 0)
+	iter, err := pdb.GetIterator()
+	if err != nil {
+		log.Printf("Error while getting PreCommitInfo DB iterator")
+		return
+	}
+	var hashKey [BlockHashSize]byte
+	preCommitHierarchy := new(PreCommitHierarchy)
+	iter.New(nil, nil)
+	for iter.Next() {
+		copy(hashKey[:], iter.Key()[:BlockHashSize])
+		preCommitInfo[hashKey] = make(map[[BlockHashSize]byte]bool, 0)
+		preCommitHierarchy.SetBytes(iter.Value())
+		for _, child := range preCommitHierarchy.childrenBlockHashes {
+			preCommitInfo[hashKey][child] = true
+		}
+	}
+	iter.Release()
+
+	if err := iter.Error(); err != nil {
+		log.Printf("Error while iterate PreCommitInfo DB")
+		return make(PreCommitInfo, 0)
+	}
+	return
+}
